+++ cpp/tests/search/search_struct_test.cpp
@@ -19,10 +19,12 @@
 #include <cudf_test/column_wrapper.hpp>
 #include <cudf_test/type_lists.hpp>
 
-#include <cudf/fixed_point/fixed_point.hpp>
+//#include <cudf/detail/iterator.cuh>
+//#include <cudf/lists/lists_column_view.hpp>
 #include <cudf/search.hpp>
+#include <cudf/table/table_view.hpp>
 
-struct DictionarySearchTest : public cudf::test::BaseFixture {
+struct StructSearchTest : public cudf::test::BaseFixture {
 };
 
 using cudf::numeric_scalar;
@@ -30,7 +32,7 @@ using cudf::size_type;
 using cudf::string_scalar;
 using cudf::test::fixed_width_column_wrapper;
 
-TEST_F(DictionarySearchTest, search_dictionary)
+TEST_F(StructSearchTest, search_dictionary)
 {
   cudf::test::dictionary_column_wrapper<std::string> input(
     {"", "", "10", "10", "20", "20", "30", "40"}, {0, 0, 1, 1, 1, 1, 1, 1});
@@ -52,7 +54,7 @@ TEST_F(DictionarySearchTest, search_dictionary)
   CUDF_TEST_EXPECT_COLUMNS_EQUAL(*result, expect_lower);
 }
 
-TEST_F(DictionarySearchTest, search_table_dictionary)
+TEST_F(StructSearchTest, search_table_dictionary)
 {
   fixed_width_column_wrapper<int32_t> column_0{{10, 10, 20, 20, 20, 20, 20, 20, 20, 50, 30},
                                                {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0}};
@@ -81,7 +83,7 @@ TEST_F(DictionarySearchTest, search_table_dictionary)
   CUDF_TEST_EXPECT_COLUMNS_EQUAL(*result, expect_upper);
 }
 
-TEST_F(DictionarySearchTest, contains_dictionary)
+TEST_F(StructSearchTest, contains_dictionary)
 {
   cudf::test::dictionary_column_wrapper<std::string> column(
     {"00", "00", "17", "17", "23", "23", "29"});
@@ -94,7 +96,7 @@ TEST_F(DictionarySearchTest, contains_dictionary)
   CUDF_TEST_EXPECT_COLUMNS_EQUAL(*result, expect);
 }
 
-TEST_F(DictionarySearchTest, contains_nullable_dictionary)
+TEST_F(StructSearchTest, contains_nullable_dictionary)
 {
   cudf::test::dictionary_column_wrapper<int64_t> column({0, 0, 17, 17, 23, 23, 29},
                                                         {1, 0, 1, 1, 1, 1, 1});
